+++ src/colors_styles.jl
@@ -281,7 +281,9 @@ end
         min = 0.0,
         easingfunction = easingflat)
 
-Calculate a value between 0 and 1 for a `point` relative to a circular area defined by `focus` and `radius`. The value will approach `max` (1.0) at the center of the circular area, and `min` (0.0) at the circumference.
+Calculate a value between 0 and 1 for a `point` relative to a circular area
+defined by `focus` and `radius`. The value will approach `max` (1.0) at the
+center of the circular area, and `min` (0.0) at the circumference.
 """
 function mask(point::Point, focus::Point, radius;
         max = 1.0,
@@ -290,6 +292,34 @@ function mask(point::Point, focus::Point, radius;
     angle = slope(focus, point)
     d = distance(focus, point)
     dref = distance(focus, focus + polar(radius, angle))
+    if d < dref
+        k = rescale(d, 0.0, dref, max, min)
+        ratio = easingfunction(k, 0.0, 1.0, 1.0)
+    else
+        ratio = 0.0
+    end
+    return ratio
+end
+
+"""
+    mask(point::Point, focus::Point, width, height)
+        max = 1.0,
+        min = 0.0,
+        easingfunction = easingflat)
+
+Calculate a value between 0 and 1 for a `point` relative to a rectangular
+area defined by `focus`, `width`, and `height`. The value will approach
+`max` (1.0) at the center, and `min` (0.0) at the edges.
+"""
+function mask(point::Point, focus::Point, width, height;
+    max = 1.0,
+    min = 0.0,
+    easingfunction = easingflat)
+
+    bb = BoundingBox(box(focus, width, height, :none))
+    ptcross = pointcrossesboundingbox(point, bb)
+    d = distance(focus, point)
+    dref = distance(focus, ptcross)
     if d < dref
         k = rescale(d, 0.0, dref, max, min)
         ratio = easingfunction(k, 0.0, 1.0, 1.0)